refactor: share shell page-load types and unify workspace tab UI
Extract AppPageLoadEvent/page copy into lib/shell so weather/young stop importing workspace; invert publications read + catalog metadata schemas into feature lib; add shared WorkspaceTaskToolbar/EmptyState; shrink e2e probeOnly and drop the empty integration probeOnly branch.
