+++ src/features/workspace/components/HomeworksTabToolbar.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
-import Plus from "@lucide/svelte/icons/plus";
 import type { HomeworkFilter } from "@/features/workspace/lib/workspace-controller-types";
-import { Button } from "$lib/components/ui/button/index.js";
-import * as ToggleGroup from "$lib/components/ui/toggle-group/index.js";
+import WorkspaceTaskToolbar from "./WorkspaceTaskToolbar.svelte";
 
 export let homeworksCopy: Record<string, string>;
 export let homeworkFilter: HomeworkFilter;
@@ -10,47 +8,14 @@ export let onHomeworkFilterChange: (value: HomeworkFilter) => void;
 export let openCreateHomeworkDialog: () => void;
 </script>
 
-<div class="grid grid-cols-[minmax(0,1fr)_auto] items-start gap-2 md:grid-cols-[minmax(0,1fr)_auto] md:items-center">
-  <div class="min-w-0 md:flex md:flex-wrap md:items-center md:gap-2 md:justify-start">
-    <ToggleGroup.Root
-      aria-label={homeworksCopy.title}
-      class="w-full min-w-0 md:w-fit"
-      type="single"
-      value={homeworkFilter}
-      variant="outline"
-      onValueChange={(value) => {
-        if (
-          value === "incomplete" ||
-          value === "completed" ||
-          value === "all"
-        ) {
-          onHomeworkFilterChange(value);
-          return;
-        }
-        onHomeworkFilterChange(homeworkFilter);
-      }}
-    >
-      <ToggleGroup.Item class="h-11 min-w-0 flex-1 text-xs md:h-8 md:flex-none md:text-sm" value="incomplete">
-        {homeworksCopy.filterIncomplete}
-      </ToggleGroup.Item>
-      <ToggleGroup.Item class="h-11 min-w-0 flex-1 text-xs md:h-8 md:flex-none md:text-sm" value="completed">
-        {homeworksCopy.filterCompleted}
-      </ToggleGroup.Item>
-      <ToggleGroup.Item class="h-11 min-w-0 flex-1 text-xs md:h-8 md:flex-none md:text-sm" value="all">
-        {homeworksCopy.filterAll}
-      </ToggleGroup.Item>
-    </ToggleGroup.Root>
-  </div>
-  <div class="flex items-center gap-2 md:justify-end">
-    <Button
-      aria-label={homeworksCopy.addButton}
-      class="size-11 md:h-8 md:w-auto md:min-w-28"
-      data-testid="workspace-homeworks-add"
-      type="button"
-      onclick={openCreateHomeworkDialog}
-    >
-      <Plus class="md:hidden" data-icon="inline-start" />
-      <span class="hidden md:inline">{homeworksCopy.addButton}</span>
-    </Button>
-  </div>
-</div>
+<WorkspaceTaskToolbar
+  ariaLabel={homeworksCopy.title}
+  filter={homeworkFilter}
+  onFilterChange={onHomeworkFilterChange}
+  filterIncompleteLabel={homeworksCopy.filterIncomplete}
+  filterCompletedLabel={homeworksCopy.filterCompleted}
+  filterAllLabel={homeworksCopy.filterAll}
+  addButtonLabel={homeworksCopy.addButton}
+  addTestId="workspace-homeworks-add"
+  onAdd={openCreateHomeworkDialog}
+/>
