+++ src/features/young/components/YoungEventDetailPage.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
-import type { WorkspacePageCopy } from "@/features/workspace/server/workspace-page-load-types";
+import type { AppPageCopy } from "@/lib/shell/page-copy";
 import type { YoungEventDetail } from "@/features/young/server/young-event-service";
 import PageLayout from "$lib/components/PageLayout.svelte";
 import Panel from "$lib/components/Panel.svelte";
 import { Button } from "$lib/components/ui/button/index.js";
 
 type Props = {
-  copy: WorkspacePageCopy;
+  copy: AppPageCopy;
   event: YoungEventDetail;
 };
