+++ tests/e2e/src/app/_shared/api-contract.ts
@@ -32,7 +32,6 @@ const probeOnlyRoutes = new Set([
   "/api/admin/users/[id]",
   "/api/auth/oauth2/device-authorization",
   "/api/auth/oauth2/token",
-  "/api/auth/.well-known/openid-configuration",
   "/api/catalog/bus",
   "/api/workspace/bus-preferences",
   "/api/workspace/subscriptions",
@@ -40,16 +39,12 @@ const probeOnlyRoutes = new Set([
   "/api/workspace/subscriptions/import-codes",
   "/api/workspace/link-pins",
   "/api/catalog/links/resolve",
-  "/api/health",
   "/api/community/section-homeworks",
   "/api/workspace/homeworks/completions",
   "/api/community/section-homeworks/[id]",
   "/api/workspace/homeworks/[id]/completion",
   "/api/mcp",
-  "/api/mcp/.well-known/oauth-authorization-server",
-  "/api/mcp/.well-known/openid-configuration",
   "/api/account/profile",
-  "/api/community/users/[identifier]",
   "/api/workspace/overview",
   "/api/workspace/homeworks",
   "/api/workspace/schedules",
